backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.11.7
pydantic-settings==2.8.1
httpx==0.28.1
pillow==10.4.0
python-multipart==0.0.12
openai>=1.30.0
python-dotenv==1.0.1
asyncpg==0.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
supabase
slowapi==0.1.9
arq==0.26.1
redis==5.0.8
